scripts: make generate_git do something
Generate a git tree from the ltsi queue of patches, by
only using git commands.

Support a quasi intelligent auto resume feature in the case
where a patch fails to apply, by using the upstream ID (if
present) to determine the current patch used for the HEAD
commit.  If that fails, use a diffstat comparison to try
and find the current patch.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>